+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 
 # Changelog
+## Version 1.0.3 on 29 May 2024
+- Fixed customer error with an empty first name and last name.
+- Fixed the email sending error when checkout as a guest and the email does not exist.
+
 ## Version 1.0.2 on 28 May 2024
 - Removed the return type hint.
 - Fixed the error from undefined variables.